improve react-router usage, forbid cross-feature imports and barrel files
