+++ docs/project-structure.md
@@ -9,7 +9,7 @@ src
 |
 +-- components        # shared components used across the entire application
 |
-+-- config            # all the global configuration, env variables etc. get exported from here and used in the app
++-- config            # global configurations, exported env variables etc.
 |
 +-- features          # feature based modules
 |
@@ -19,7 +19,7 @@ src
 |
 +-- providers         # all of the application providers
 |
-+-- routes            # routes configuration
++-- routes            # routes of the application
 |
 +-- stores            # global state stores
 |
@@ -45,57 +45,43 @@ src/features/awesome-feature
 |
 +-- hooks       # hooks scoped to a specific feature
 |
-+-- routes      # route components for a specific feature pages
-|
 +-- stores      # state stores for a specific feature
 |
 +-- types       # typescript types for TS specific feature domain
 |
 +-- utils       # utility functions for a specific feature
-|
-+-- index.ts    # entry point for the feature, it should serve as the public API of the given feature and exports everything that should be used outside the feature
 ```
 
-Everything from a feature should be exported from the `index.ts` file which behaves as the public API of the feature.
+Do not import across the features. Instead, compose different features at the application level in the `routes` folder. This way, you can ensure that each feature is independent and can be easily moved or removed without affecting other parts of the application.
 
-You should import stuff from other features only by using:
-
-`import {AwesomeComponent} from "@/features/awesome-feature"`
-
-and not
-
-`import {AwesomeComponent} from "@/features/awesome-feature/components/awesome-component`
-
-This can also be configured in the ESLint configuration to disallow the later import by the following rule:
+To forbid cross-feature imports, you can use ESLint:
 
 ```js
-{
-    rules: {
-        'no-restricted-imports': [
-            'error',
+'import/no-restricted-paths': [
+    'error',
+    {
+        zones: [
             {
-                patterns: ['@/features/*/*'],
+                target: './src/features/comments',
+                from: './src/features',
+                except: ['./comments'],
+            },
+            {
+                target: './src/features/discussions',
+                from: './src/features',
+                except: ['./discussions'],
+            },
+            {
+                target: './src/features/teams',
+                from: './src/features',
+                except: ['./teams'],
+            },
+            {
+                target: './src/features/users',
+                from: './src/features',
+                except: ['./users'],
             },
         ],
-
-    // ...rest of the configuration
-}
+    },
+],
 ```
-
-To prevent circular dependencies, here is another ESLint rule that can be used:
-
-```js
-{
-    rules: {
-        'import/no-cycle': 'error',
-    // ...rest of the configuration
-}
-```
-
-This was inspired by how [NX](https://nx.dev/) handles libraries that are isolated but available to be used by the other modules. Think of a feature as a library or a module that is self-contained but can expose different parts to other features via its entry point. This approach would also make it easier to split the application in a monorepo in the future.
-
-This way, you can ensure that the codebase is clean and easy to maintain.
-
-If you are still getting circular dependencies, consider breaking the feature into smaller features or moving the shared code to the `lib` folder. Or you can always opt-out of the barrel export pattern and import the files directly.
-
-Sometimes, it will make more sense to move the whole API layer to the `lib` folder, especially if it is shared across multiple features. The same goes for the `stores` folder. If you have a global store that is used across the entire application, it should be moved to the `stores` folder.
