--- src/features/auth/api/get-user.ts
@@ -1,7 +1,0 @@
-import { api } from '@/lib/api-client';
-
-import { AuthUser } from '../types';
-
-export const getUser = (): Promise<AuthUser> => {
-  return api.get('/auth/me');
-};
