--- src/features/auth/api/logout.ts
@@ -1,5 +1,0 @@
-import { api } from '@/lib/api-client';
-
-export const logout = (): Promise<void> => {
-  return api.post('/auth/logout');
-};
