--- src/features/auth/api/register.ts
@@ -1,34 +1,0 @@
-import { z } from 'zod';
-
-import { api } from '@/lib/api-client';
-
-import { UserResponse } from '../types';
-
-export const registerInputSchema = z
-  .object({
-    email: z.string().min(1, 'Required'),
-    firstName: z.string().min(1, 'Required'),
-    lastName: z.string().min(1, 'Required'),
-    password: z.string().min(1, 'Required'),
-  })
-  .and(
-    z
-      .object({
-        teamId: z.string().min(1, 'Required'),
-        teamName: z.null().default(null),
-      })
-      .or(
-        z.object({
-          teamName: z.string().min(1, 'Required'),
-          teamId: z.null().default(null),
-        }),
-      ),
-  );
-
-export type RegisterInput = z.infer<typeof registerInputSchema>;
-
-export const registerWithEmailAndPassword = (
-  data: RegisterInput,
-): Promise<UserResponse> => {
-  return api.post('/auth/register', data);
-};
