--- src/features/auth/index.ts
@@ -1,8 +1,0 @@
-export * from './lib/auth';
-export * from './lib/authorization';
-export * from './lib/protected-route';
-
-export * from './routes/register';
-export * from './routes/login';
-
-export * from './types';
