--- src/features/auth/lib/protected-route.tsx
@@ -1,19 +1,0 @@
-import { Navigate, useLocation } from 'react-router-dom';
-
-import { useUser } from './auth';
-
-export const ProtectedRoute = ({ children }: { children: React.ReactNode }) => {
-  const user = useUser();
-  const location = useLocation();
-
-  if (!user.data) {
-    return (
-      <Navigate
-        to={`/auth/login?redirectTo=${encodeURIComponent(location.pathname)}`}
-        replace
-      />
-    );
-  }
-
-  return children;
-};
