--- src/features/auth/types/index.ts
@@ -1,13 +1,0 @@
-export type AuthUser = {
-  id: string;
-  email: string;
-  firstName: string;
-  lastName: string;
-  bio: string;
-  role: 'ADMIN' | 'USER';
-};
-
-export type UserResponse = {
-  jwt: string;
-  user: AuthUser;
-};
