+++ src/features/discussions/components/create-discussion.tsx
@@ -2,7 +2,7 @@ import { Plus } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
 import { Form, FormDrawer, Input, Textarea } from '@/components/ui/form';
-import { Authorization, ROLES } from '@/features/auth';
+import { Authorization, ROLES } from '@/lib/authorization';
 import { useNotifications } from '@/stores/notifications';
 
 import {
