+++ src/features/discussions/components/delete-discussion.tsx
@@ -2,7 +2,7 @@ import { Trash } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
 import { ConfirmationDialog } from '@/components/ui/dialog';
-import { Authorization, ROLES } from '@/features/auth';
+import { Authorization, ROLES } from '@/lib/authorization';
 import { useNotifications } from '@/stores/notifications';
 
 import { useDeleteDiscussion } from '../api/delete-discussion';
