--- src/features/discussions/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './routes/discussion';
-export * from './routes/discussions';
