+++ src/features/discussions/types/index.ts
@@ -1,9 +1,15 @@
-import { User } from '@/features/users';
-import { BaseEntity } from '@/types';
+import { Entity } from '@/types';
 
-export type Discussion = {
+export type Discussion = Entity<{
   title: string;
   body: string;
   teamId: string;
-  author: User;
-} & BaseEntity;
+  author: Entity<{
+    firstName: string;
+    lastName: string;
+    email: string;
+    role: 'ADMIN' | 'USER';
+    teamId: string;
+    bio: string;
+  }>;
+}>;
