--- src/features/teams/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-
-export * from './api/get-teams';
