+++ src/features/users/components/update-profile.tsx
@@ -2,7 +2,7 @@ import { Pen } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
 import { Form, FormDrawer, Input, Textarea } from '@/components/ui/form';
-import { useUser } from '@/features/auth';
+import { useUser } from '@/lib/auth';
 import { useNotifications } from '@/stores/notifications';
 
 import {
