--- src/features/users/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-
-export * from './routes/users';
-export * from './routes/profile';
