+++ src/lib/react-query.ts
@@ -1,4 +1,8 @@
-import { UseMutationOptions, DefaultOptions } from '@tanstack/react-query';
+import {
+  UseMutationOptions,
+  DefaultOptions,
+  QueryClient,
+} from '@tanstack/react-query';
 
 export const queryConfig = {
   queries: {
@@ -7,6 +11,10 @@ export const queryConfig = {
     retry: false,
   },
 } satisfies DefaultOptions;
+
+export const queryClient = new QueryClient({
+  defaultOptions: queryConfig,
+});
 
 export type ApiFnReturnType<FnType extends (...args: any) => Promise<any>> =
   Awaited<ReturnType<FnType>>;
