+++ src/routes/index.tsx
@@ -1,45 +1,9 @@
 import { Suspense } from 'react';
-import { Routes, Route, Outlet } from 'react-router-dom';
+import { Outlet, createBrowserRouter } from 'react-router-dom';
 
 import { DashboardLayout } from '@/components/layouts';
 import { Spinner } from '@/components/ui/spinner';
-import { ProtectedRoute } from '@/features/auth';
-import { lazyImport } from '@/utils/lazy-import';
-
-const { ProfileRoute } = lazyImport(
-  () => import('@/features/users'),
-  'ProfileRoute',
-);
-const { UsersRoute } = lazyImport(
-  () => import('@/features/users'),
-  'UsersRoute',
-);
-const { RegisterRoute } = lazyImport(
-  () => import('@/features/auth'),
-  'RegisterRoute',
-);
-const { LoginRoute } = lazyImport(
-  () => import('@/features/auth'),
-  'LoginRoute',
-);
-const { DiscussionRoute } = lazyImport(
-  () => import('@/features/discussions'),
-  'DiscussionRoute',
-);
-const { DiscussionsRoute } = lazyImport(
-  () => import('@/features/discussions'),
-  'DiscussionsRoute',
-);
-
-const { DashboardRoute } = lazyImport(
-  () => import('./dashboard'),
-  'DashboardRoute',
-);
-const { LandingRoute } = lazyImport(() => import('./landing'), 'LandingRoute');
-const { NotFoundRoute } = lazyImport(
-  () => import('./not-found'),
-  'NotFoundRoute',
-);
+import { ProtectedRoute } from '@/lib/auth';
 
 const MainApp = () => {
   return (
@@ -57,30 +21,82 @@ const MainApp = () => {
   );
 };
 
-export const AppRoutes = () => {
-  return (
-    <Routes>
-      <Route path="/" element={<LandingRoute />} />
-      <Route path="/auth/register" element={<RegisterRoute />} />
-      <Route path="/auth/login" element={<LoginRoute />} />
-      <Route
-        path="/app"
-        element={
-          <ProtectedRoute>
-            <MainApp />
-          </ProtectedRoute>
-        }
-      >
-        <Route path="/app/discussions" element={<DiscussionsRoute />} />
-        <Route
-          path="/app/discussions/:discussionId"
-          element={<DiscussionRoute />}
-        />
-        <Route path="/app/users" element={<UsersRoute />} />
-        <Route path="/app/profile" element={<ProfileRoute />} />
-        <Route path="/app/" element={<DashboardRoute />} />
-      </Route>
-      <Route path="*" element={<NotFoundRoute />} />
-    </Routes>
-  );
-};
+export const router = createBrowserRouter([
+  {
+    path: '/',
+    lazy: async () => {
+      const { LandingRoute } = await import('./landing');
+      return { Component: LandingRoute };
+    },
+  },
+  {
+    path: '/auth/register',
+    lazy: async () => {
+      const { RegisterRoute } = await import('./auth/register');
+      return { Component: RegisterRoute };
+    },
+  },
+  {
+    path: '/auth/login',
+    lazy: async () => {
+      const { LoginRoute } = await import('./auth/login');
+      return { Component: LoginRoute };
+    },
+  },
+  {
+    path: '/app',
+    element: (
+      <ProtectedRoute>
+        <MainApp />
+      </ProtectedRoute>
+    ),
+    children: [
+      {
+        path: 'discussions',
+        lazy: async () => {
+          const { DiscussionsRoute } = await import(
+            '@/routes/app/discussions/discussions'
+          );
+          return { Component: DiscussionsRoute };
+        },
+      },
+      {
+        path: 'discussions/:discussionId',
+        lazy: async () => {
+          const { DiscussionRoute } = await import(
+            '@/routes/app/discussions/discussion'
+          );
+          return { Component: DiscussionRoute };
+        },
+      },
+      {
+        path: 'users',
+        lazy: async () => {
+          const { UsersRoute } = await import('./app/users');
+          return { Component: UsersRoute };
+        },
+      },
+      {
+        path: 'profile',
+        lazy: async () => {
+          const { ProfileRoute } = await import('./app/profile');
+          return { Component: ProfileRoute };
+        },
+      },
+      {
+        path: '',
+        lazy: async () => {
+          const { DashboardRoute } = await import('./app/dashboard');
+          return { Component: DashboardRoute };
+        },
+      },
+    ],
+  },
+  {
+    path: '*',
+    lazy: async () => {
+      const { NotFoundRoute } = await import('./not-found');
+      return { Component: NotFoundRoute };
+    },
+  },
+]);
