+++ src/types/index.ts
@@ -2,3 +2,7 @@ export type BaseEntity = {
   id: string;
   createdAt: number;
 };
+
+export type Entity<T> = {
+  [K in keyof T]: T[K];
+} & BaseEntity;
